Move out background, controls from MapViewport
